MouseTracking in ComponentDidMount()
